requirements.txt
fastapi==0.115.6
starlette==0.41.3
uvicorn[standard]==0.34.0
pydantic==2.10.5
sqlmodel==0.0.22
SQLAlchemy==2.0.37
concurrent-log-handler==0.9.25
PyYAML==6.0.2
PyMySQL==1.1.1
aiomysql==0.2.0
PyJWT==2.10.1
aiohttp==3.11.11
aioredis[hiredis]==2.0.1
elasticsearch[async]~=8.11.1
sentry-sdk==2.20.0
python-multipart==0.0.20
apscheduler==3.11.0
uvloop==0.21.0 # 仅支持linux
# dev
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1